updatertest.ApplyUpdates
