+++ updater/dockerurl/dependencies_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/thepwagner/action-update-go/updater"
-	"github.com/thepwagner/action-update-go/updater/dockerurl"
 	"github.com/thepwagner/action-update-go/updatertest"
 )
 
@@ -18,6 +17,5 @@ func TestUpdater_Dependencies(t *testing.T) {
 		},
 	}
 
-	updaterFactory := func(root string) updater.Updater { return dockerurl.NewUpdater(root) }
-	updatertest.DependenciesFixtures(t, updaterFactory, cases)
+	updatertest.DependenciesFixtures(t, updaterFactory(), cases)
 }
