+++ updater/dockerurl/updater_test.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 
 	"github.com/thepwagner/action-update-go/updater"
+	"github.com/thepwagner/action-update-go/updater/dockerurl"
+	"github.com/thepwagner/action-update-go/updatertest"
 )
 
 //go:generate mockery --outpkg dockerurl_test --output . --testonly --name repoClient --structname mockRepoClient --filename mockrepoclient_test.go
@@ -21,3 +23,7 @@ var (
 	dep     = updater.Dependency{Path: depPath, Version: previousVersion}
 	update  = updater.Update{Path: depPath, Previous: previousVersion, Next: nextVersion}
 )
+
+func updaterFactory(opts ...dockerurl.UpdaterOpt) updatertest.Factory {
+	return func(root string) updater.Updater { return dockerurl.NewUpdater(root, opts...) }
+}
