+++ updater/gomod/updater_test.go
@@ -10,6 +10,10 @@ import (
 	"github.com/thepwagner/action-update-go/updatertest"
 )
 
+func updaterFactory(opts ...gomod.UpdaterOpt) updatertest.Factory {
+	return func(root string) updater.Updater { return gomod.NewUpdater(root, opts...) }
+}
+
 func updaterFromFixture(t *testing.T, fixture string, opts ...gomod.UpdaterOpt) *gomod.Updater {
 	tempDir := updatertest.TempDirFromFixture(t, fixture)
 	return gomod.NewUpdater(tempDir, opts...)
