--- updatertest/applyupdate.go
@@ -1,16 +1,0 @@
-package updatertest
-
-import (
-	"fmt"
-	"testing"
-
-	deepcopy "github.com/otiai10/copy"
-	"github.com/stretchr/testify/require"
-)
-
-func TempDirFromFixture(t *testing.T, fixture string) string {
-	tempDir := t.TempDir()
-	err := deepcopy.Copy(fmt.Sprintf("testdata/%s", fixture), tempDir)
-	require.NoError(t, err)
-	return tempDir
-}
